+++ example/app.js
@@ -5,9 +5,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 injectTapEventPlugin();
-import {Dialog, TextField, Styles} from 'material-ui';
-let ThemeManager = Styles.ThemeManager;
-import {Tabs, Tab} from '../lib/index.js';
+import {Dialog, FlatButton, TextField} from 'material-ui';
+import {Tabs, Tab} from '../src/index.js';
 
 import DynamicTabContent from './DynamicTabContent';
 import DynamicTabBadge from './DynamicTabBadge';
@@ -44,26 +43,26 @@ class App extends React.Component {
 
     this.state = {
       tabs:[
-        (<Tab key={'tab0'} title={'unclosable tab'} disableClose={true} >
+        (<Tab key={'tab0'} title={'unclosable tab'} disableClose={true} {...this.makeListeners('tab0')}>
           <div>
             <h1>This tab cannot close</h1>
           </div>
         </Tab>),
-        (<Tab key={'tab1'} title={'1stTab'} beforeTitle={icon} >
+        (<Tab key={'tab1'} title={'1stTab'} beforeTitle={icon} {...this.makeListeners('tab1')}>
           <div>
             <h1>This is tab1</h1>
           </div>
         </Tab>),
-        (<Tab key={'tab2'} title={'2ndTab Too long Toooooooooooooooooo long'} beforeTitle={fonticon} >
+        (<Tab key={'tab2'} title={'2ndTab Too long Toooooooooooooooooo long'} beforeTitle={fonticon} {...this.makeListeners('tab2')}>
           <div>
             <pre>Lorem ipsum dolor sit amet, consectetur adipisicing elit,
             </pre>
           </div>
         </Tab>),
-        (<Tab key={'tab3'} title={'Dynamic tab'} afterTitle={badge}>
+        (<Tab key={'tab3'} title={'Dynamic tab'} afterTitle={badge} {...this.makeListeners('tab3')}>
           <DynamicTabContent/>
         </Tab>),
-        (<Tab key={'tab4'} title={'Custom container'} containerStyle={{backgroundColor: 'gray', width: '50%'}}>
+        (<Tab key={'tab4'} title={'Custom container'} containerStyle={{backgroundColor: 'gray', width: '50%'}} {...this.makeListeners('tab4')}>
           <div>
             <h1>This is tab4 with custom container style</h1>
           </div>
@@ -79,6 +78,15 @@ class App extends React.Component {
   //   };
   // }
 
+  makeListeners(key){
+    return {
+      onClick: (e) => { console.log('onClick', key, e);}, // never called
+      onContextMenu: (e) => { console.log('onContextMenu', key, e);},
+      onDoubleClick: (e) => { console.log('onDoubleClick', key, e); this.handleTabDoubleClick(key, e)},
+      onMouseEnter: (e) => { console.log('onMouseEnter', key, e);},
+      onMouseLeave: (e) => { console.log('onMouseLeave', key, e);}
+    }
+  }
 
   handleTabSelect(e, key, currentTabs) {
     console.log('handleTabSelect key:', key);
@@ -98,7 +106,7 @@ class App extends React.Component {
   handleTabAddButtonClick(e, currentTabs) {
     // key must be unique
     const key = 'newTab_' + Date.now();
-    let newTab = (<Tab key={key} title='untitled' >
+    let newTab = (<Tab key={key} title='untitled' {...this.makeListeners(key)}>
                     <div>
                       <h1>New Empty Tab</h1>
                     </div>
@@ -111,24 +119,11 @@ class App extends React.Component {
     });
   }
 
-  handleTabDoubleClick(e, key) {
-
-    let tab = _.find(this.state.tabs, (t) => {
-      return t.key === key;
-    });
+  handleTabDoubleClick(key, e) {
     this.setState({
-      editTabKey: key
-    }, () => {
-      this.refs.dialog.show();
+      editTabKey: key,
+      dialogOpen: true
     });
-  }
-
-  handleTabMouseEnter(e, key) {
-    console.log('tab mouseEnter key:', key);
-  }
-
-  handleTabMouseLeave(e, key) {
-    console.log('tab mouseLeave key:', key);
   }
 
   _onDialogSubmit() {
@@ -140,13 +135,16 @@ class App extends React.Component {
         return tab;
       }
     });
-    this.setState({tabs: newTabs}, () => {
-      this.refs.dialog.dismiss();
+    this.setState({
+      tabs: newTabs,
+      dialogOpen: false
     });
   }
 
   _onDialogCancel() {
-    this.refs.dialog.dismiss();
+    this.setState({
+      dialogOpen: false
+    })
   }
 
   _onDialogShow() {
@@ -167,8 +165,17 @@ class App extends React.Component {
   render() {
 
     let standardActions = [
-      { text: 'Cancel', onClick: this._onDialogCancel.bind(this) },
-      { text: 'Submit', onClick: this._onDialogSubmit.bind(this), ref: 'submit' }
+      <FlatButton
+        label="Cancel"
+        secondary={true}
+        onTouchTap={this._onDialogCancel.bind(this)}
+      />,
+      <FlatButton
+        label="Submit"
+        primary={true}
+        keyboardFocused={true}
+        onTouchTap={this._onDialogSubmit.bind(this)}
+      />
     ];
 
 
@@ -182,9 +189,6 @@ class App extends React.Component {
           onTabClose={this.handleTabClose.bind(this)}
           onTabAddButtonClick={this.handleTabAddButtonClick.bind(this)}
           onTabPositionChange={this.handleTabPositionChange.bind(this)}
-          onTabDoubleClick={this.handleTabDoubleClick.bind(this)}
-          onTabMouseEnter={this.handleTabMouseEnter.bind(this)}
-          onTabMouseLeave={this.handleTabMouseLeave.bind(this)}
           tabs={this.state.tabs}
           shortCutKeys={
             {
@@ -200,8 +204,8 @@ class App extends React.Component {
           title="Change tab name"
           ref="dialog"
           actions={standardActions}
-          actionFocus="submit"
           modal={true}
+          open={this.state.dialogOpen}
           onShow={this._onDialogShow.bind(this)}>
           <TextField
             ref='input' style={{width: '90%'}}/>