+++ karma.conf.js
@@ -10,7 +10,7 @@ module.exports = function(config) {
 
     // frameworks to use
     // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-    frameworks: ['browserify','mocha'],
+    frameworks: ['browserify', 'mocha'],
 
 
     // list of files / patterns to load in the browser
@@ -38,6 +38,7 @@ module.exports = function(config) {
           'babelify', {
             'presets': [
               'es2015',
+              'stage-2',
               'react'
             ]
           }